Change logic to show description also for active filters whose facet is defined several times on the page.
At the moment it can happen if one facet is defined several times on the page, the description of that facet, will be overwritten with the last defined facet description.
Example: The facet property is defined two times as color and size. The description of the active filters will be always size because it is the last defined one. This happen because we use the facet name "property" as id to match the description with the correct facet.